Implemented Yousef's Part
